Adiciona funcionalidade de buscar uma turma por id
